+++ morpheus-graphql-examples-servant/src/Server/API/Simple.hs
@@ -11,6 +11,7 @@
 
 module Server.API.Simple
   ( simpleApi,
+    rootResolver,
   )
 where