+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.13.1 - Unreleased Changes
+
+### new features
+
+- `Data.Morpheus.Document` exposes `RootResolverConstraint`
+- `Data.Morpheus.Server` exposes `httpPlayground`
+
 ## 0.13.0 - 22.06.2020
 
 ### breaking changes